+++ crates/holochain/src/core/workflow/call_zome_workflow.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use super::app_validation_workflow;
 use super::app_validation_workflow::AppValidationError;
 use super::app_validation_workflow::Outcome;
@@ -21,9 +19,11 @@ use holochain_keystore::MetaLairClient;
 use holochain_p2p::HolochainP2pDna;
 use holochain_state::host_fn_workspace::SourceChainWorkspace;
 use holochain_state::source_chain::SourceChainError;
-use holochain_zome_types::record::Record;
-
 use holochain_types::prelude::*;
+use holochain_zome_types::record::Record;
+use parking_lot::Mutex;
+use std::collections::HashSet;
+use std::sync::Arc;
 use tracing::instrument;
 
 #[cfg(test)]
@@ -283,10 +283,12 @@ where
                 Ok(op) => op,
                 Err(outcome_or_err) => return map_outcome(Outcome::try_from(outcome_or_err)),
             };
+            let fetched_dependencies = Arc::new(Mutex::new(HashSet::new()));
 
             let outcome = app_validation_workflow::validate_op(
                 &op,
                 workspace.clone().into(),
+                fetched_dependencies.clone(),
                 &network,
                 &ribosome,
                 &conductor_handle,
@@ -294,11 +296,47 @@ where
             .await;
             let outcome = outcome.or_else(Outcome::try_from);
             map_outcome(outcome)?;
-            chain_record = app_validation_workflow::op_to_record(op, omitted_entry);
+            chain_record = op_to_record(op, omitted_entry);
         }
     }
 
     Ok(())
+}
+
+fn op_to_record(op: Op, omitted_entry: Option<Entry>) -> Record {
+    match op {
+        Op::StoreRecord(StoreRecord { mut record }) => {
+            if let Some(e) = omitted_entry {
+                // NOTE: this is only possible in this situation because we already removed
+                // this exact entry from this Record earlier. DON'T set entries on records
+                // anywhere else without recomputing hashes and signatures!
+                record.entry = RecordEntry::Present(e);
+            }
+            record
+        }
+        Op::StoreEntry(StoreEntry { action, entry }) => {
+            Record::new(SignedActionHashed::raw_from_same_hash(action), Some(entry))
+        }
+        Op::RegisterUpdate(RegisterUpdate {
+            update, new_entry, ..
+        }) => Record::new(SignedActionHashed::raw_from_same_hash(update), new_entry),
+        Op::RegisterDelete(RegisterDelete { delete, .. }) => Record::new(
+            SignedActionHashed::raw_from_same_hash(delete),
+            omitted_entry,
+        ),
+        Op::RegisterAgentActivity(RegisterAgentActivity { action, .. }) => Record::new(
+            SignedActionHashed::raw_from_same_hash(action),
+            omitted_entry,
+        ),
+        Op::RegisterCreateLink(RegisterCreateLink { create_link, .. }) => Record::new(
+            SignedActionHashed::raw_from_same_hash(create_link),
+            omitted_entry,
+        ),
+        Op::RegisterDeleteLink(RegisterDeleteLink { delete_link, .. }) => Record::new(
+            SignedActionHashed::raw_from_same_hash(delete_link),
+            omitted_entry,
+        ),
+    }
 }
 
 fn map_outcome(